+++ docs/experiments/raven-target/QueryBindings.cs
@@ -8,6 +8,9 @@ static class QueryBindings
             public static class Enumerable {
                 public static Collections.Iterable<T> Where<T>(this Collections.Iterable<T> source, Func<T, bool> predicate) => default;
                 public static Collections.Iterable<U> Select<T,U>(this Collections.Iterable<T> source, Func<T,U> selector) => default;
+                public static Option<T> First<T>(this Collections.Iterable<T> source) => default;
+                public static Option<T> Last<T>(this Collections.Iterable<T> source) => default;
+                public static Result<T, SingleError> Single<T>(this Collections.Iterable<T> source) => default;
                 public static Collections.ArrayList<T> ToList<T>(this Collections.Iterable<T> source) => default;
             }
         }
@@ -28,6 +31,8 @@ static class QueryBindings
         var (expected, returns) = reference.Name switch {
             "Where" => (new[] { source, $"System.Func<{types[0]},Boolean>" }, source),
             "Select" => (new[] { source, $"System.Func<{types[0]},{types[1]}>" }, $"System.Collections.Iterable<{types[1]}>"),
+            "First" or "Last" => (new[] { source }, $"System.Option<{types[0]}>"),
+            "Single" => (new[] { source }, $"System.Result<{types[0]},System.Linq.SingleError>"),
             "ToList" => (new[] { source }, $"System.Collections.ArrayList<{types[0]}>"),
             _ => throw new InvalidDataException("Unsupported query operator.")
         };